Assigning lights to grid works.

+++ ConsoleProject/PointLight.cu
@@ -1,8 +1,7 @@
 #include "pch.h"
 #include "PointLight.h"
 
-
-__device__ void PointLight::Update(float dt)
+__device__ void PointLight::Update(long double dt)
 {
 	return;
 }
